+++ backend/src/main/java/com/example/Main/global/Config/WebMvcConfig.java
@@ -1,0 +1,17 @@
+package com.example.Main.global.Config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+    @Value("${custom.fileDirPath}")
+    private String fileDirPath;
+
+    public void addResourceHandlers (ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/file/**")
+                .addResourceLocations("file:///" + fileDirPath + "/");
+    }
+}
