+++ frontend/styles/components/birthday/birthday.module.scss
@@ -1,7 +1,7 @@
 @import "../../globals/font-mixin.scss";
 
 .birthdayForm {
-  margin: 20px 0;
+  
 
   .birthdayP {
     display: flex;
