+++ frontend/styles/components/profile/profile.module.scss
@@ -1,0 +1,10 @@
+.profileImageSection {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.imageControls {
+  display: flex;
+  gap: 10px;
+}
